lint: add exception to eq for empty string
The callback function `helper_operator_eq_check` for the eq operator,
has an overloaded behavior. It first tries regex matching and if that is
not successful, it tests for regular string equality checking.

This caused some issues when testing against an empty string. A check
value of empty string used to match any kind of input, which is quite
counter-intuitive.

With this change we are adding an exception for the empty string case to
force string equality check in that case. Also adding a deprecation
notice for the overloaded regex checking behavior to inform the users
that it will be removed in the future.

Works-on: 1993735
Signed-off-by: Mert Kırpıcı <mert.kirpici@canonical.com>