tests/unit: fail under %100 coverage
This was missed during the snap template update, we expect %100 test
coverage in this project.

This also revealed a sneaky bug in the process. When the
jujulint.config.Config class got initialized it was using the sys.argv
of the pytest process, causing a testing error and not being able to
pass arguments to pytest.

Signed-off-by: Mert Kırpıcı <mert.kirpici@canonical.com>