lint: add support for fetching rules over url
This patch extends the functionality of juju-lint in several ways:

- juju-lint now accepts multiple rules files to be given through the
  command line and processes them in the order that they are given to
update the rules. i.e.

    $ juju-lint -c ./rules-1.yaml -c ./rules-2.yaml ...

if the rules-2.yaml file overrides an option, in the resulting rules
dictionary, it will take precedence. Note that this is a deep update
operation, which means, it will follow through the nested dictionaries.
For implementation details, see jujulint.util.deep_update().

- juju-lint now accepts urls as rules file command line arguments. i.e.

    $ juju-lint -c https://git.launchpad.net/myrepo/plain/rules.yaml ..

the underlying implementation utilizes urllib which is part of the
python stdlib. Therefore what schemas it supports, we should be able to
support given that the resource is accessible to us.

Works-on: 1987359

Signed-off-by: Mert Kırpıcı <mert.kirpici@canonical.com>