Replace canonical-livepatch with ubuntu-advantage in all rulesets
canonical-livepatch is now deprecated, and should be replaced with
ubuntu-advantage (soon ubuntu-pro).

canonical-livepatch is retained as an optional subordinate to avoid
throwing errors about it on existing older clouds.

The absence of the ubuntu-advantage will however trigger a linting
error, regardless of whether livepatch esists in the model or not.

Closes-bug: 2032714